Fix stdlib compilation with dotty when using sbt 1.3
When we compile the standard library with dotty, we actually compile the
scala-library and dotty-library sources together in one go since we
can't in general reuse a dotty-library compiled with an old
scala-library. There were two problems with the way this was done that
were somehow masked so far (likely because of classpath pollution, but I
don't know what changed in sbt 1.3 exactly):

- Compiling the stdlib requires having all sources on the -sourcepath,
  but the sources from dotty-library itself were missing.
- the files in scalaShadowing/ in dotty-library do need to be compiled,
  the comment above the exclusion was wrong.

Additionally, I also removed the other exclusions done to files in
dotty-library because they're no longer necessary.