Reduce memory footprint of evaluated values in LazyList
The LazyList implementation uses two objects per evaluated value,
a LazyList and one State instance. So the overhead for evaluated
values is 2x compared to Stream.

This PR reduces the overhead to one LazyList instance with two
fields (head and tail) per evaluated value.

Thread safety is implemented through checking the volatile `head`
field and synchronizing on `this` for initialization. This is
the equivalent to the previous implementation which uses a `lazy val`.

Breaking changes:

Serialization is not compatible. Non-evaluated (tails of) streams are
serialized using plain object serialization. The internals are
different and serialization is incompatible in both directions.