Independent attachment to enforce Match desugaring
Follow-up for a recent change (PR 10775) where Match desugaring
is enforced (avoid emitting switches) in methods later transformed
by async.

This PR adds a separate attachment to enforce desugaring instead
of relying on the `AsyncAttachment`. That makes it easier for
compiler plugins to also enforce the desugaring (the `AsyncAttachment`
is not public, and adding it has the effect of enabling the async
transformation).