Do not expand lambdas used in super constructor calls.
That was necessary when `-Ydelambdafy:method` was introduced.
However, since then, the codegen was improved so that delambdafy
targets that do not refer to `this` are compiled as `static`.

If a lambda *actually* refers to `this`, as demonstrated in
`t10035.scala`, the previous code missed the issue. Trying to run
the code `t10035.scala` actually caused a `LinkageError`. This
commit moves it to `neg/`, which is a progression.