Select symbol overload in tree unpickling
References to external symbols are pickled as owner + name, so
overloads cannot be resolved. This only affects tree unpickling,
i.e., only annotation arguments.

At some point in history, the unpickler was using
`Infer.inferMethodAlternative`, but not anymore since fff93cd049.

This commit performs overloading resolution based on the Select
tree's type instead: when unpickling `t.f`, the `tpe` of the tree
is unpickled. This type can be used to select the correct overload.