Don't crash pickling Literal trees with a null `tpe`
Constant annotation arguments are `Literal(Constant(v))`
ASTs with a non-null `tpe` assigned by typer.

But it's relatively easy to create an untyped Literal
in a compiler plugin, the factory `Literal(Constant(v))`
returns a tree with a `null` type.

This commit prevents an NPE in pickling for this case.