fix: error location for aliased import selector
In the case of something like `import foo.{DoesntExist => Bar}`, the
error location should be on `DoesntExist`. The previous check in error
message construction for whether the name is introduced by a selector is
not quite right - really we just care about whether the name is _used_
by the selector (though in the non-rename case this is the same thing).