Make ClassfileParser robust for directly reading Java inner classes
Usually when the ClassfileParser reads a classfile for a Java inner
class, the entry point is the outer class, the inner symbol is
created in `enterOwnInnerClasses`.

But when scanning the classpath, a Symbol is also created because
of the presence of the `C$I.class` classfile. (This symbol is later
unlinked and marked invalid.) When invoking the classfile parser on
this symbol directly, reading generic signatures can fail because
they may refer to type variables defined in the outer class:

```
class C1<X> {
  // signature: Lp/C1<TX;>; -- `TX;` references outer `X`
  class I extends C1<X> { }
}
```