Completions: exclude top-level class symbols with a `$`
Completions were changed recently to include symbols with a `$` in their
name (PR 11024).

In SymbolLoaders, the compiler creates a class symbol for every `.class`
file on the classpath. This results in many class symbols that we don't
want to see in completions: inner classes, specialized classes, module
classes.

This commit excludes non-completed top-level symbols that have a `$` in
their name.